% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_es.R
\name{delta_es}
\alias{delta_es}
\title{Calculate change in whole-body energy storage}
\usage{
delta_es(
  fm_start,
  fm_end,
  ffm_start,
  ffm_end,
  fm_units = "kg",
  ffm_units = fm_units,
  start_date,
  end_date,
  n_days = 1,
  ...
)
}
\arguments{
\item{fm_start}{starting fat mass}

\item{fm_end}{ending fat mass}

\item{ffm_start}{starting fat-free mass}

\item{ffm_end}{ending fat-free mass}

\item{fm_units}{character scalar indicating units of the values passed for \code{fm_start} and \code{fm_end}}

\item{ffm_units}{character scalar indicating units of the values passed for \code{ffm_start} and \code{ffm_end}}

\item{start_date}{the start date of the measurement period}

\item{end_date}{the end date of the measurement period}

\item{n_days}{optional The number of days in the measurement period (will be
calculated from \code{start_date} and \code{end_date} if provided.
Otherwise the default is 1, such that calculations will reflect raw change
in energy storage)}

\item{...}{arguments passed to \code{\link[anthropometry]{get_es}}}
}
\value{
numeric vector of change values
}
\description{
Calculate change in whole-body energy storage
}
\examples{
delta_es(50.5, 50.7, 75.1, 74.9, n_days = 14)
}
