% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntakeBalance_function.R
\name{IntakeBalance-function}
\alias{IntakeBalance-function}
\alias{IntakeBalance}
\alias{IntakeBalance_default}
\alias{IntakeBalance_df}
\title{Calculate energy intake over time via the intake-balance method}
\usage{
IntakeBalance(
  fm_start,
  fm_end,
  ffm_start,
  ffm_end,
  fm_units = "kg",
  ffm_units = fm_units,
  ee_per_day,
  ee_units = "kcal",
  df = NULL,
  ...
)

IntakeBalance_default(
  fm_start,
  fm_end,
  ffm_start,
  ffm_end,
  fm_units = "kg",
  ffm_units = fm_units,
  ee_per_day,
  ee_units = "kcal",
  ...
)

IntakeBalance_df(
  df,
  fm_start,
  fm_end,
  ffm_start,
  ffm_end,
  fm_units = "kg",
  ffm_units = fm_units,
  ee_per_day,
  ee_units = "kcal",
  start_date = NULL,
  end_date = NULL,
  n_days = NULL,
  ...,
  es_name = "delta_ES",
  ei_name = "EI"
)
}
\arguments{
\item{fm_start}{starting fat mass}

\item{fm_end}{ending fat mass}

\item{ffm_start}{starting fat-free mass}

\item{ffm_end}{ending fat-free mass}

\item{fm_units}{character scalar indicating units of the values passed for \code{fm_start} and \code{fm_end}}

\item{ffm_units}{character scalar indicating units of the values passed for \code{ffm_start} and \code{ffm_end}}

\item{ee_per_day}{energy expenditure expressed in units per day}

\item{ee_units}{character scalar indicating units of the energy expenditure values}

\item{df}{optional. A data frame on which to operate}

\item{...}{Arguments passed to \code{\link{delta_es}}}

\item{start_date}{the start date of the measurement period}

\item{end_date}{the end date of the measurement period}

\item{n_days}{optional The number of days in the measurement period (will be
calculated from \code{start_date} and \code{end_date} if provided.
Otherwise the default is 1, such that calculations will reflect raw change
in energy storage)}

\item{es_name}{character scalar. Name to assign the new variable in \code{df}
containing information about change in energy storage}

\item{ei_name}{character scalar. Name to assign to the new variable in
\code{df} containing information about energy intake}
}
\value{
Either a numeric vector of energy intake values (if \code{df = NULL},
  or a data frame appended with energy intake data
}
\description{
Calculate energy intake over time via the intake-balance method
}
\details{
This function operates in different ways, depending on whether anything is
passed for the \code{df} argument. If not, the arguments are passed into
\code{IntakeBalance_default}, where the expectation is that numeric values
are passed for fat mass, fat-free mass, and energy expenditure variables.
Otherwise, so, the arguments are passed into \code{IntakeBalance_df}, where
the expectation is that character scalars are passed, which represent the
column names of variables in \code{df} which contain the corresponding
information.
}
\examples{

d <- data.frame(
  fm_start = c(50.5, 70.2), fm_end = c(50.7, 70.0),
  ffm_start = c(75.1, 90.3), ffm_end = c(74.9, 90.3),
  ee = c(1950, 2473), n_days = c(14, 14)
)

IntakeBalance(50.5, 50.7, 75.1, 74.9, ee_per_day = 1950, n_days = 14)

IntakeBalance(
  "fm_start", "fm_end", "ffm_start", "ffm_end",
  ee_per_day = "ee", n_days = "n_days", df = d
)

}
